Pull immutable jmap wrapper out into utils
